@@ -13,14 +13,19 @@
         "@popperjs/core": "^2.10.2",
         "axios": "^0.25",
         "bootstrap": "^5.1.3",
+        "file-loader": "^6.2.0",
         "laravel-mix": "^6.0.6",
         "lodash": "^4.17.19",
         "postcss": "^8.1.14",
         "resolve-url-loader": "^3.1.2",
         "sass": "^1.32.11",
         "sass-loader": "^11.0.1",
-        "vue": "^2.6.12",
-        "vue-loader": "^15.9.7",
+        "vue": "^3.0.0",
+        "vue-loader": "^16.0.0-rc.1",
         "vue-template-compiler": "^2.6.12"
+    },
+    "dependencies": {
+        "vue-router": "^4.0.14",
+        "vuex": "^4.0.2"
     }
 }
