@@ -5,22 +5,14 @@
  */
 
 require('./bootstrap');
-
 window.Vue = require('vue').default;
-
-/**
- * The following block of code may be used to automatically register your
- * Vue components. It will recursively scan this directory for the Vue
- * components and automatically register them with their "basename".
- *
- * Eg. ./components/ExampleComponent.vue -> <example-component></example-component>
- */
 
 // const files = require.context('./', true, /\.vue$/i)
 // files.keys().map(key => Vue.component(key.split('/').pop().split('.')[0], files(key).default))
 
-Vue.component('example-component', require('./components/ExampleComponent.vue').default);
-Vue.component('test', require('./components/test.vue').default);
+import {createApp} from 'vue';
+import App from "./components/App"
+import store from "./store"
 
 /**
  * Next, we will create a fresh Vue application instance and attach it to
@@ -28,6 +20,12 @@ Vue.component('test', require('./components/test.vue').default);
  * or customize the JavaScript scaffolding to fit your unique needs.
  */
 
-const app = new Vue({
-    el: '#app',
-});
+
+
+// app.config.unwrapInjectedRef = true
+
+const app = createApp(App)
+
+app.use(store);
+app.mount("#app")
+
